backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
loguru==0.7.3
pydantic==2.11.7
bcrypt==4.3.0
python_jose==3.4.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
pytest==8.4.0
pydantic[email]
python-multipart==0.0.20
ruff==0.12.0
httpx==0.28.1
typing_extensions>=4.0.0
requests==2.32.4
tqdm==4.67.1
APScheduler==3.11.0